+++ app/lib/ui/components/button/button_demo_screen.dart
@@ -120,6 +120,11 @@ class _ButtonDemoState extends State<_ButtonDemo> {
       themeController?.setOnColoredSurface(customizationState?.hasOnColoredBox);
     });
 
+    // Adding post-frame callback to update theme based on customization state
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      themeController?.setOnBorderRadiusButtonState(customizationState?.hasRoundedCorner);
+    });
+
     if (customizationState?.hasOnColoredBox == true) {
       return OudsColoredBox(
         color: customizationState?.hasOnColoredBox == true ? OudsColoredBoxColor.brandPrimary : OudsColoredBoxColor.statusNeutralMuted,
@@ -183,8 +188,15 @@ class _CustomizationContentState extends State<_CustomizationContent> {
     return CustomizableSection(
       children: [
         CustomizableSwitch(
+          title: context.l10n.app_components_common_roundedCorner_label,
+          value: customizationState!.hasRoundedCorner,
+          onChanged: (value) {
+            customizationState.hasRoundedCorner = value;
+          },
+        ),
+        CustomizableSwitch(
           title: context.l10n.app_common_enabled_label,
-          value: customizationState!.hasEnabled,
+          value: customizationState.hasEnabled,
           onChanged:
 
               /// Specific case: Enabled disabled if style is 'Loading'